Extract visible multi-agent tmux wake helpers
